+++ src/sisl/tests/test_geometry.py
@@ -1453,6 +1453,7 @@ class TestGeometry:
         from_ase = gr.new(to_ase)
         assert gr.equal(from_ase, R=False)
 
+    @pytest.mark.xfail(reason="pymatgen backconversion sets nsc=[3, 3, 3], we need to figure this out")
     def test_geometry_pymatgen_to(self):
         pytest.importorskip("pymatgen", reason="pymatgen not available")
         gr = sisl_geom.graphene()