+++ src/sisl/unit/siesta.py
@@ -9,6 +9,7 @@ The conversion factors are taken directly from Siesta
 which means these unit conversions should be used for Siesta "stuff".
 """
 
+from sisl._environ import get_environ_variable, register_environ_variable
 from sisl._internal import set_module
 
 from .base import UnitParser
@@ -20,19 +21,25 @@ from .base import unit_table
 __all__  = ["unit_group", "unit_convert", "unit_default", "units"]
 
 
-unit_table_siesta = dict({key: dict(values) for key, values in unit_table.items()})
+register_environ_variable("SISL_UNIT_SIESTA", "codata2018",
+                          "Choose default units used when parsing Siesta files. [codata2018, legacy]",
+                          process=str.lower)
 
-unit_table_siesta["length"].update({
+
+unit_table_siesta_codata2018 = dict({key: dict(values) for key, values in unit_table.items()})
+unit_table_siesta_legacy = dict({key: dict(values) for key, values in unit_table.items()})
+
+unit_table_siesta_legacy["length"].update({
     "Bohr": 0.529177e-10,
 })
 
-unit_table_siesta["time"].update({
+unit_table_siesta_legacy["time"].update({
     "mins": 60.,
     "hours": 3600.,
     "days": 86400.,
 })
 
-unit_table_siesta["energy"].update({
+unit_table_siesta_legacy["energy"].update({
     "meV": 1.60219e-22,
     "eV": 1.60219e-19,
     "mRy": 2.17991e-21,
@@ -49,7 +56,7 @@ unit_table_siesta["energy"].update({
     "cm^-1": 1.986e-23,
 })
 
-unit_table_siesta["force"].update({
+unit_table_siesta_legacy["force"].update({
     "eV/Ang": 1.60219e-9,
     "eV/Bohr": 1.60219e-9*0.529177,
     "Ry/Bohr": 4.11943e-8,
@@ -57,28 +64,31 @@ unit_table_siesta["force"].update({
 })
 
 
+# Check for the correct handlers
+_def_unit = get_environ_variable("SISL_UNIT_SIESTA")
+if _def_unit in ("codata2018", "codata"):
+    unit_table_siesta = unit_table_siesta_codata2018
+elif _def_unit in ("legacy", "original"):
+    unit_table_siesta = unit_table_siesta_legacy
+else:
+    raise ValueError(f"Could not understand SISL_UNIT_SIESTA={_def_unit}, expected one of [codata2018, legacy]")
+
+
 @set_module("sisl.unit.siesta")
-def unit_group(unit, tbl=None):
-    global unit_table_siesta
-    if tbl is None:
-        return u_group(unit, unit_table_siesta)
+def unit_group(unit, tbl=unit_table_siesta):
     return u_group(unit, tbl)
 
 
 @set_module("sisl.unit.siesta")
-def unit_default(group, tbl=None):
-    global unit_table_siesta
-    if tbl is None:
-        return u_default(group, unit_table_siesta)
+def unit_default(group, tbl=unit_table_siesta):
     return u_default(group, tbl)
 
 
 @set_module("sisl.unit.siesta")
-def unit_convert(fr, to, opts=None, tbl=None):
-    global unit_table_siesta
-    if tbl is None:
-        return u_convert(fr, to, opts, unit_table_siesta)
+def unit_convert(fr, to, opts=None, tbl=unit_table_siesta):
     return u_convert(fr, to, opts, tbl)
+
 
 # create unit-parser
 units = UnitParser(unit_table_siesta)
+units_legacy = UnitParser(unit_table_siesta_legacy)